style: input 요소를 줄에 맞춰 정렬
