+++ src/components/InputSection/CardCVCNumberInputSection.tsx
@@ -2,7 +2,7 @@ import useCardCVCNumber from '../../hooks/useCardCVCNumber';
 import ErrorMessage from '../ErrorMessage';
 import InputField from '../InputField';
 import InputSection from './InputSection';
-
+import styles from './inputSection.module.css';
 
 const CardCVCNumberInputSection = () => {
   const { cardCVCNumber, setCardCVCNumber, isError, errorMessage } =
@@ -10,11 +10,13 @@ const CardCVCNumberInputSection = () => {
 
   return (
     <InputSection title="CVC 번호를 입력해주세요" description="" subtitle="CVC">
-      <InputField
-        value={cardCVCNumber}
-        onChange={setCardCVCNumber}
-        isError={isError}
-      ></InputField>
+      <div className={styles.inputContainer}>
+        <InputField
+          value={cardCVCNumber}
+          onChange={setCardCVCNumber}
+          isError={isError}
+        ></InputField>
+      </div>
 
       <ErrorMessage message={errorMessage} />
     </InputSection>
